apirest_sm/requirements.txt
aniso8601==8.1.1
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
passlib==1.7.4
pep517==0.8.2
progress==1.5
PyJWT==1.7.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.23
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
